chore: cleanup project and removed storybook
