--- .storybook/.babelrc
@@ -1,8 +1,0 @@
-{
-  "plugins": ["react-require"],
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ]
-}
