+++ src/components/Header/MobileMenu.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { Grid, Box, Flex, Collapse } from "@chakra-ui/react";
 import { useDisclosure } from "@chakra-ui/hooks";
-import { Button } from "@components/ui";
+import Button from "@components/Button";
 import MenuToggle from "./MenuToggle";
 import { navigation } from "./constants";
