--- src/components/layouts/index.ts
@@ -1,4 +1,0 @@
-export { default as AnimateOnScreen } from "./AnimateOnScreen";
-export { default as Container } from "./Container";
-export { MotionBox, LazyMotionBox } from "./MotionBox";
-export { default as PageLayout } from "./PageLayout";
