--- src/components/sections/index.tsx
@@ -1,1 +1,0 @@
-export { default as Header } from "./Header";
