--- src/components/ui/Link/NextLink.tsx
@@ -1,5 +1,0 @@
-import Link from "next/link";
-
-const NextLink = Link;
-
-export default NextLink;
