+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
 
-import theme from "@config/theme";
+import theme from "../theme";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
