--- src/services/index.ts
@@ -1,1 +1,0 @@
-export { default as api } from "./api";
