+++ src/theme/colors.ts
@@ -1,0 +1,8 @@
+export const colors = {
+  primary: "#1C987D",
+  secondary: "#31D8B4",
+  light: "#F3F4F8",
+  black: "#1C1C1C",
+  white: "#fff",
+  gray: "#9B9FA4",
+};
