+++ src/theme/fonts.ts
@@ -1,0 +1,4 @@
+export const fonts = {
+  heading: "Open Sans",
+  body: "Open Sans",
+};
