+++ src/theme/styles.ts
@@ -1,0 +1,16 @@
+export const styles = {
+  global: {
+    html: {
+      scrollBehavior: "smooth",
+    },
+    "::-webkit-scrollbar-track": {
+      background: "primary",
+    },
+    "::-webkit-scrollbar": {
+      width: "6px",
+    },
+    "::-webkit-scrollbar-thumb": {
+      background: "secondary",
+    },
+  },
+};
