+++ tsconfig.json
@@ -1,11 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": false,
@@ -19,39 +15,18 @@
     "jsx": "preserve",
     "baseUrl": ".",
     "paths": {
-      "@utils": [
-        "src/utils/index"
-      ],
-      "@services": [
-        "src/services/index"
-      ],
-      "@mocks/*": [
-        "src/mocks/*"
-      ],
-      "@hooks/*": [
-        "src/hooks/*"
-      ],
-      "@i18n/*": [
-        "src/i18n/*"
-      ],
-      "@config/*": [
-        "src/config/*"
-      ],
-      "@components/*": [
-        "src/components/*"
-      ],
-      "@images/*": [
-        "public/images/*"
-      ]
+      "@utils": ["src/utils/index"],
+      "@services": ["src/services/index"],
+      "@mocks/*": ["src/mocks/*"],
+      "@hooks/*": ["src/hooks/*"],
+      "@i18n/*": ["src/i18n/*"],
+      "@config/*": ["src/config/*"],
+      "@theme/*": ["src/theme/*"],
+      "@components/*": ["src/components/*"],
+      "@images/*": ["public/images/*"]
     },
     "incremental": true
   },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "exclude": ["node_modules"]
 }
